SageMaker/text_preprocessing/T5_model/requirements.txt
openpyxl
nltk
gensim
janome
python-docx